PO/docker/downloader/app/requirements.txt
pika
pymysql
requests
python-dotenv
